Add history filter for place history
