+++ src/history/pages/GroupHistory.vue
@@ -1,18 +1,8 @@
 <template>
   <div>
-    <div
-      class="row no-wrap items-center justify-between bg-white q-px-sm q-py-xs"
-    >
-      <QSelect
-        v-model="typus"
-        :options="typusOptions"
-        emit-value
-        map-options
-        outlined
-        hide-bottom-space
-        dense
-      />
-    </div>
+    <HistoryFilter
+      v-model="typus"
+    />
     <QCard
       class="no-mobile-margin no-shadow grey-border"
     >
@@ -34,69 +24,17 @@
 </template>
 
 <script setup>
-import { QSelect, QCard } from 'quasar'
-import { computed } from 'vue'
-import { useI18n } from 'vue-i18n'
+import { QCard } from 'quasar'
+import { ref } from 'vue'
 
 import { useCurrentGroupService } from '@/group/services'
 import { useHistoryListQuery } from '@/history/queries'
-import { useQueryParams } from '@/utils/mixins/bindRoute'
 
+import HistoryFilter from '@/history/components/HistoryFilter'
 import HistoryList from '@/history/components/HistoryList'
 import RandomArt from '@/utils/components/RandomArt'
 
-const { t } = useI18n()
-
-const { typus } = useQueryParams({
-  typus: null,
-})
-
-const typusOptions = computed(() => [
-  {
-    label: t('HISTORY_FILTER.ALL'),
-    value: null,
-  },
-  ...[
-    'GROUP_CREATE',
-    'GROUP_MODIFY',
-    'GROUP_JOIN',
-    'GROUP_LEAVE',
-    'STORE_CREATE',
-    'STORE_MODIFY',
-    'STORE_DELETE',
-    'ACTIVITY_CREATE',
-    'ACTIVITY_MODIFY',
-    'ACTIVITY_DELETE',
-    'SERIES_CREATE',
-    'SERIES_MODIFY',
-    'SERIES_DELETE',
-    'ACTIVITY_DONE',
-    'ACTIVITY_JOIN',
-    'ACTIVITY_LEAVE',
-    'ACTIVITY_MISSED',
-    'APPLICATION_DECLINED',
-    'MEMBER_BECAME_EDITOR',
-    'ACTIVITY_DISABLE',
-    'ACTIVITY_ENABLE',
-    'GROUP_LEAVE_INACTIVE',
-    'GROUP_CHANGE_PHOTO',
-    'GROUP_DELETE_PHOTO',
-    'MEMBER_REMOVED',
-    'ACTIVITY_TYPE_CREATE',
-    'ACTIVITY_TYPE_MODIFY',
-    'ACTIVITY_TYPE_DELETE',
-    'USER_LOST_EDITOR_ROLE',
-    'PLACE_TYPE_CREATE',
-    'PLACE_TYPE_MODIFY',
-    'PLACE_TYPE_DELETE',
-    'AGREEMENT_CREATE',
-    'AGREEMENT_MODIFY',
-    'MEMBER_GOT_ROLE',
-  ].map(typus => ({
-    label: t(`HISTORY_FILTER.${typus}`),
-    value: typus,
-  })),
-])
+const typus = ref(null)
 
 const { groupId } = useCurrentGroupService()
 const {
